Adding activity table.

+++ src/main/java/pl/edu/uam/rest/marsey/resources/AbstractCandidatesResource.java
@@ -1,9 +1,13 @@
 package pl.edu.uam.rest.marsey.resources;
 
 import com.wordnik.swagger.annotations.ApiOperation;
+import com.wordnik.swagger.annotations.ApiResponse;
+import com.wordnik.swagger.annotations.ApiResponses;
+import org.apache.commons.lang3.math.NumberUtils;
 import org.apache.log4j.Logger;
 import pl.edu.uam.rest.marsey.db.MarseyDatabase;
-import pl.edu.uam.rest.marsey.exceptions.CandidateException;
+import pl.edu.uam.rest.marsey.exceptions.CandidateNotFoundException;
+import pl.edu.uam.rest.marsey.exceptions.NumericException;
 import pl.edu.uam.rest.marsey.model.Candidate;
 
 import javax.ws.rs.*;
@@ -34,7 +38,7 @@ public abstract class AbstractCandidatesResource {
     @Produces(MediaType.APPLICATION_JSON)
     public Response createCandidate(Candidate candidate) {
         Candidate dbCandidate = new Candidate("", candidate.getName(), candidate.getSurname(), candidate.getSex(),
-                candidate.getOccupation(), candidate.getAge(), candidate.getHeight());
+                candidate.getOccupation(), candidate.getAge(), candidate.getHeight(), candidate.getFitness());
 
         Candidate createdCandidate = getDatabase().createCandidate(dbCandidate);
 
@@ -44,10 +48,17 @@ public abstract class AbstractCandidatesResource {
 
     @GET
     @Path("/{candidateId}")
-    @ApiOperation(value = "Get candidate by id", response = Candidate.class, position = 3)
+    @ApiOperation(value = "Find candidate by id", response = Candidate.class, position = 3)
     @Produces(MediaType.APPLICATION_JSON)
     public Candidate getCandidate(@PathParam("candidateId") String candidateId) throws Exception {
         Candidate candidate = getDatabase().getCandidate(candidateId);
+        
+        logger.info(NumberUtils.isDigits(candidateId));
+
+        if (!NumberUtils.isDigits(candidateId) || Integer.valueOf(candidateId) < 0) {
+            logger.error("Invalid id provided.");
+            throw new NumericException("Invalid id", "Podano nieprawidłowe id");
+        }
 
         if (candidateId.equals("db")) {
             logger.error("Database error");
@@ -56,8 +67,7 @@ public abstract class AbstractCandidatesResource {
 
         if (candidate == null) {
             logger.error("Candidate not found");
-            throw new CandidateException("Candidate not found", "Kandydat nie został znaleziony", 
-                    "http://docu.pl/errors/user-not-found");
+            throw new CandidateNotFoundException("Candidate not found", "Kandydat nie został znaleziony");
         }
 
         return candidate;
@@ -69,16 +79,23 @@ public abstract class AbstractCandidatesResource {
     @ApiOperation(value = "Update candidate", notes = "update candidate info", response = Candidate.class, position = 4)
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
+    @ApiResponses(value = {
+            @ApiResponse(code = 400, message = "Invalid ID supplied"),
+            @ApiResponse(code = 404, message = "Candidate not found")
+    })
     public Response updateCandidate(@PathParam("candidateId") String candidateId, Candidate candidate) {
-        
+        int id = Integer.valueOf(candidateId);
         logger.info("PUT candidate" + candidate);
+
+        if (!NumberUtils.isNumber(candidateId) || Integer.valueOf(candidateId) < 0) {
+            logger.error("Invalid id provided.");
+            throw new NumericException("Invalid id", "Podano nieprawidłowe id");
+        }
 
         Candidate candidateFromDb = getDatabase().getCandidate(candidateId);
         if (candidateFromDb == null) {
             logger.error("Candidate not found");
-            throw new CandidateException("Candidate not found",
-                    "Kandydat nie został znaleziony",
-                    "http://docu.pl/errors/user-not-found");
+            throw new CandidateNotFoundException("Candidate not found", "Kandydat nie został znaleziony");
         }
 
         Candidate updatedCandidate = getDatabase().updateCandidate(candidateId, candidate);
@@ -92,13 +109,17 @@ public abstract class AbstractCandidatesResource {
     @Path("/{candidateId}")
     @ApiOperation(value = "Remove a candidate", notes = "Delete a candidate", position = 5)
     public Response removeCandidate(@PathParam("candidateId") String candidateId) {
-        
+
+        if (!NumberUtils.isNumber(candidateId) || Integer.valueOf(candidateId) < 0) {
+            logger.error("Invalid id provided.");
+            throw new NumericException("Invalid id", "Podano nieprawidłowe id");
+        }
+
         Candidate candidateFromDb = getDatabase().getCandidate(candidateId);
         if (candidateFromDb == null) {
             logger.error("Candidate not found");
-            throw new CandidateException("Candidate not found",
-                    "Kandydat nie został znaleziony",
-                    "http://docu.pl/errors/user-not-found");
+            throw new CandidateNotFoundException("Candidate not found",
+                    "Kandydat nie został znaleziony");
         }
 
         getDatabase().removeCandidate(candidateId);
